Annotation update

+++ annotations/nocedit.exe/pseudocode/src/core/crossbow.cpp/CCrossbow_renderOpaque_FUN_00448e00.cpp
@@ -23,7 +23,7 @@ int __cdecl core_crossbow_cpp_CCrossbow_renderOpaque_FUN_00448e00(CCrossbow *thi
     return 0;
   }
   render_flags = -NAN;
-  animation_frame = (float)core_crossbow_cpp_FUN_00449540();
+  animation_frame = core_crossbow_cpp_FUN_00449540();
   core_dmodel_cpp_CKeyFramedModelInstance_prepareForRendering_FUN_00478d20
             (&(this_ptr->base).model,animation_frame,(int)render_flags);
   core_actor_cpp_CDemonActor_restoreRenderState_FUN_00408b40((CDemonActor *)this_ptr);
